feat: Update pipelines table and filter (#319)
