+++ src/constants/pipelineTypes.ts
@@ -4,5 +4,4 @@ export enum PIPELINE_TYPES {
   REVIEW = 'review',
   DEPLOY = 'deploy',
   CLEAN = 'clean',
-  AUTOTEST_RUNNER = 'autotestRunner',
 }
