--- src/k8s/groups/Tekton/PipelineRun/utils/createAutotestRunnerPipelineRunInstance/index.test.ts
@@ -1,2 +1,0 @@
-// Todo: write new mocks
-test('true', () => {});
