+++ src/pages/pipelines/page.tsx
@@ -1,10 +1,7 @@
 import React from 'react';
 import { PageLogicWrapper } from '../../components/PageLogicWrapper';
 import { DialogContextProvider } from '../../providers/Dialog';
-import { FilterContextProvider } from '../../providers/Filter';
 import { ResourceActionListContextProvider } from '../../providers/ResourceActionList';
-import { getDefaultNamespace } from '../../utils/getDefaultNamespace';
-import { matchFunctions } from './constants';
 import { PermissionsContextProvider } from './providers/Permissions/provider';
 import { PageView } from './view';
 
@@ -13,15 +10,9 @@ export default function () {
     <PageLogicWrapper>
       <PermissionsContextProvider>
         <DialogContextProvider>
-          <FilterContextProvider
-            entityID={`PIPELINE_RUN_LIST_OVERVIEW::${getDefaultNamespace()}`}
-            matchFunctions={matchFunctions}
-            saveToLocalStorage
-          >
-            <ResourceActionListContextProvider>
-              <PageView />
-            </ResourceActionListContextProvider>
-          </FilterContextProvider>
+          <ResourceActionListContextProvider>
+            <PageView />
+          </ResourceActionListContextProvider>
         </DialogContextProvider>
       </PermissionsContextProvider>
     </PageLogicWrapper>
